Tweak thresholds for statistic assertions
I've been running this spec in a row to make sure that there are no
flakiness, and after quite a while I got a failed spec in a case where
79%, rather than >= 80% of samples where delivered to the correct
thread.

To avoid this showing up in CI, I've decided to relax further the
assertion. We also have the sanity checks to validate that we get
enough samples, so we can be a bit more lenient on the actual ratio.