[PROF-6691] Send span id in profiles as number and not string
**What does this PR do?**:

This PR changes the `span id` sent in profiling data (and used to
power the "Code Hotspots" tab when viewing a trace) as a number in
the underlying pprof, instead of as a string.

**Motivation**:

The numeric representation is smaller and also avoids needing to use
the string table to represent this information.

We've wanted to do this change for a while, but only now was the
profiling backend extended to support this.

**Additional Notes**:

The `local root span id` is still recorded as a string because
there's a feature in libdatadog that relies/assumes this string
representation.

I plan to later change libdatadog at a later time to allow us to
send both values as a number.

**How to test the change?**:

Change includes code coverage. Furthermore, you can validate
this is working correctly by profiling an application that is
generating traces, and checking that the "Code Hotspots" for
spans show up correctly.