Bootstrap support for using stdatomic.h header
Because this header was only added for C11, we need to validate that
it's available and usable in the current system.

This header was available as far back as Ubuntu 16.04 so I'm going
to require it for now; if we later learn that some Linux version we
really want to support is missing, we can add fallback support for
older atomic primitives.