[PROF-6557] Record stats for time spent sampling in new profiler
**What does this PR do?**:

This PR adds tracking of the min, max and total wall-time that the
profiler spent sampling.

For now this is exposed as `#stats` and not particularly recorded
or sent anywhere.

**Motivation**:

I'm preparing to introduce dynamic sampling rate into the profiler,
and as part of it, I'll need to use the time the profiler spent
sampling to decide how often the profiler runs.

For a long time I've wanted to add these stats for debugging and
analysis so I decided to do a small PR to add them first, and then
later have the dynamic sampling rate PR rely on them as well.

**Additional Notes**:

(N/A)

**How to test the change?**:

Change includes code coverage.