Refactor: Calculate `sampling_time_ns` in `CpuAndWallTimeWorker`
Originally the calculation happened in
`cpu_and_wall_time_collector_sample` because that's where we were
getting the "before" timestamp.

Now that the "before" timestamp comes from the `CpuAndWallTimeWorker`,
we can move the whole calculation there. This will also be usedful
for implementing the dynamic sampling rate, since we need the
"after" timestamp for it.