Introduce helpers for min and max
It's so bizarre that C doesn't have this as a built-in in the
standard library, especially when you can write pages and pages on
stack overflow and compiler docs on how the many ways people try to
do this common operation can lead to issues with double evaluation,
shadowing, and whatnot.

Ever heard of show, don't tell, C devs? ;P