[PROF-6557] Implement dynamic sampling rate for new profiler
**What does this PR do?**:

This PR adds a dynamic sampling rate implementation for the new
profiler. The old profiler implementation already included a similar
strategy, and it has been serving us quite well.

To more easily test and document the new feature, even though in the
end it's not that much code, I chose to implement it in a separate
module -- `Collectors::DynamicSamplingRate`.

This module gets called by the `CpuAndWallTimeWorker` at key times
during sampling and in the loop that triggers sampling, so that
its internal tracking gets taken into account.

**Motivation**:

Our profilers get deployed in quite unpredictable situations in terms of
system resources. While they can provide key information to help
customers solve their performance problems, the profilers must always
be careful not to make performance problems worse. This is where the
idea of a dynamic sampling rate comes in.

Instead of sampling at a fixed sample rate, the actual sampling rate
should be decided by also observing the impact that running the
profiler is having. This protects against issues such as the profiler
being deployed in very busy machines or containers with unrealistic
CPU restrictions.

**Additional Notes**:

See the comments on `collectors_dynamic_sampling_rate.c` for a deeper
explanation of how this works

**How to test the change?**:

This change includes code coverage. Furthermore, profiles should
still continue to be correctly taken for applications.