Inline `REPORT_ISSUE` into its user
The `REPORT_ISSUE` string is somewhat specific to the error that was
using it, because it mentions "if you needed to use this", where this
is the env var to disable building the native extension.

So to avoid confusion, I decided to inline it on the check it belongs
to.