[PROF-6900] Fix wrong libdatadog version being picked during profiler build
**What does this PR do?**:

This PR fixes an issue I spotted recently relating to the libdatadog
version that gets picked during profiler build.

Specifically, I had the following setup:

* Two versions of the libdatadog gem locally installed
  (1.0.1.1.0, 0.9.0.1.0)
* I tried to install ddtrace 1.8.0 from rubygems

The ddtrace 1.8.0 release has a dependency on libdatadog 0.9.x.

But, what I observed is that building the profiler (and thus installing
ddtrace) failed because the profiling native extension was trying to
compile and link to libdatadog 1.0.x instead of 0.9.x.

Upon investigation, I've discovered that the issue is that the
`extconf.rb` script (which gets called by rubygems during installation)
is not run with the equivalent of `bundle exec`, e.g. it can access all
gems (and versions) that are installed locally, not just the ones that
are defined in the dependencies of our gem in the `gemspec`.

Thus, at least on my machine, when we did `require 'libdatadog'`
to access the `libdatadog` gem, rubygems picked a version which did
not match the one that was stated in the `gemspec`.

To fix this, I've modified the `native_extension_helpers.rb`
(which gets called by the `extconf.rb`) to tell RubyGems to activate
the exact version of `libdatadog` that we need.

Additionally, I've added error handling so we gracefully recover
from any errors that can happen on this step.

**Motivation**:

Avoid impacting customers that are installing ddtrace (and avoid
having them endup having to disable profiling to install ddtrace).

**Additional Notes**:

Hopefully no customers ever got bitten by this (we had no reports thus
far). To reproduce this, they'd need to have a weird combination of
having a more modern version of libdatadog than they needed, and then
install an older version of ddtrace.

**How to test the change?**:

See instructions above for the setup I had that triggered this
issue.