RFC: Move `agent.port` back to core
**What does this PR do?**:

This PR is a follow-up to #2494 and undoes some of the changes done in
that PR.

In particular, it moves the `option :port` definition back to the
core settings.

**Motivation**:

The `port` gets used by components other than tracing (both
profiling and telemetry), as an input to the `AgentSettingsResolver`.

We've discussed this internally and this RFC PR is a potential
path forward for making sure all components can still use the `port`
in the future.

**Additional Notes**:

This is not just a `git revert` -- I've tried to minimize the diff
from #2494.

**How to test the change?**:

Change includes test coverage.