Expose development docker container ports to localhost only
**What does this PR do?**:

This PR modifies our `docker-compose.yml` files to bind any ports
on the development docker containers only for access by the host machine.

This should not impact any development use-cases.

**Motivation**:

If the ports are not bound to localhost, they could potentially be
accessed by other machines in the same network, which is not a great
idea.

**Additional Notes**:

(N/A)

**How to test the change?**:

Check that running tests using `docker-compose` still works, as well
as running them from your development machine directly.