[PROF-6660] Add `profiling_enabled` state to environment logger output
**What does this PR do?**:

This PR adds a `profiling_enabled` entry in the environment logger
output, which reflects the current configuration for profiling.

**Motivation**:

This was requested internally as being useful to support, and at least
Node and Java already provide this key.

We also had a TODO for it, so one less TODO.

**Additional Notes**:

I hesitated between adding the configuration of the profiler being
on, or the profiler actually being on (e.g. it's possible that
the profiler has been requested on, but can't be enabled because
for instance google-protobuf is missing).

My thinking is that the information about if the profiler really
is on could be instead relayed via telemetry, so it may not
be worth duplicating it here.

**How to test the change?**:

Change includes test coverage.