Document how to publish updates to test container images
**What does this PR do?**:

This PR adds a new section in the container images `README.md`
explaining the steps to publish new versions of the images, as well
as a quick note to every `Dockerfile` to remind people that the note in
the `README.md` exists.

**Motivation**:

Recently a change was merged, but not published, and I was a bit
confused on how to actually publish the images, so I made a note to come
back and document this, so hopefully this is clear for others and for a
next time.

**Additional Notes**:

(N/A)

**How to test the change?**:

Test the steps I've documented!