Drop experimental TruffleRuby testing scaffolding
**What does this PR do?**:

This PR removes our half-working scaffolding for testing dd-trace-rb
with TruffleRuby.

We currently are not prioritizing getting this over the finish line,
so this is not really working, but its existence adds complexity and
toil (e.g. trying to maintain the Dockerfile).

The intention here is that we'll re-add this once we're ready to get
TruffleRuby fully running in CI.

**Motivation**:

Reduce CI complexity and maintenance toil.

**Additional Notes**:

I'm still a big fan of TruffleRuby! I think it's going to be the
future of Ruby ;)

**How to test the change?**:

Validate that CI is still green.