Test only with JRuby 9.2.21.0 in CI (disable 9.2.8.0)
**What does this PR do?**:

This PR disables testing of JRuby 9.2.8.0 in CI, as well as
appraisal generation for it (as well as removes the existing
appraisals).

**Motivation**:

Maintaining JRuby 9.2.8.0 working has been quite a burden: this version
has many, many, many known issues -- you will notice that there have
been **13** other bugfix releases on top of it; latest JRuby is
9.2.31.0.

Some examples of PRs where this JRuby version caused toil are
 #2565, #2558, #2448, #2412, #1869, etc

We've decided to stop running it in CI for now, and plan to evaluate
if we should remove it entirely from our test harness.

We recommend that all our customers use JRuby 9.2.21.0, and also
recommend evaluating the move to JRuby 9.3, as the JRuby developers have
marked JRuby 9.2 as end-of-life:
<https://twitter.com/jruby/status/1611063274459566106>.

**Additional Notes**:

I didn't have to remove the appraisal gemfiles from the tree, but
since I disabled the tasks that automatically keep them up-to-date,
they wouldn't be maintained anyway.

If/when we do want to revert this change, we should regenerate them using
the usual rake tasks.

**How to test the change?**:

Validate that CI is green and that the JRuby 9.2.8.0 configuration is
not shown there.