Add appraisals in preparation for Roda support
**What does this PR do?**:

This PR was extracted from the Roda integration PR (#2368) and includes
**only** the addition of `roda` to the appraisals + the result of running
`bundle exec rake appraisal:install`.

**Motivation**:

I decided to break this from #2368 as fiddling with appraisals has
caused a lot of back-and-forth and lost time on a PR that is otherwise
in good shape.

Let's merge this as fast as possible, and then that PR can focus only
on adding Roda.

**Additional Notes**:

(N/A)

**How to test the change?**:

Roda testing will come as part of #2368. This PR is mostly a no-op, and CI
should still be green.