Minor: Import upstream safety check into the profiler custom `rb_profile_frames`
**What does this PR do?**:

This PR imports the fix in https://github.com/ruby/ruby/pull/7116 into
the custom version of `rb_profile_frames` that we ship inside ddtrace.

**Motivation**:

As I mention in the inline comments, I don't expect this extra safety
check to ever be triggered, but just in case I decided to add it
anyway, since the cost of getting my expectation wrong is a VM crash.

**Additional Notes**:

N/A

**How to test the change?**:

Adding test coverage for this is... difficult. Our current test of
this logic is indirect (e.g. via our own code for sampling), and
reproducing the state to trigger this is hard.

So for now I only added the check without extra testing.