[PROF-7145] Remove support for profiling Ruby 2.2
**What does this PR do?**:

This PR removes support for profiling Ruby 2.2 using the Continuous
Profiling product. Other products shipped in ddtrace (tracing,
application security monitoring, CI, etc) are unaffected.

This dropping of support does not impact our customer's ability to
install the latest dd-trace-rb on Ruby 2.2. What it does is

1. Skip compilation of the profiling native extension on Ruby 2.2
2. Print a warning message when customers try to enable profiling on
   Ruby 2.2 (but does not otherwise block or break their application)

The warning message shown to customers is:

> W, [2023-02-01T11:42:57.022293 #410738]  WARN -- ddtrace: [ddtrace]
> Profiling was requested but is not supported, profiling disabled:
> Your ddtrace installation is missing support for the Continuous
> Profiler because the profiler only supports Ruby 2.3 or newer.
> Upgrade to a modern Ruby to enable profiling for your app.

Customers are welcome to continue using older versions of dd-trace-rb
to profile their Ruby 2.2 (as well as 2.1) applications.

(This PR is similar to #2140 where we dropped support for Ruby 2.1)

**Motivation**:

There is little customer interest on profiling Ruby 2.2, and
supporting old Rubies also comes with an extra tax -- just look
how much code and complexity is being removed by this PR.

**Additional Notes**:

(N/A)

**How to test the change?**:

* Validate that test suite still runs successfully on Ruby 2.2
* Validate that you get the above error message when trying to
  profile a Ruby 2.2 application