Make sorbet happy
Unfortunately sorbet started complaining about

```
lib/datadog/core/configuration/agent_settings_resolver.rb:32: Method Datadog::Core::Configuration::AgentSettingsResolver::AgentSettings#initialize redefined with argument adapter as a keyword argument https://srb.help/4010
    32 |            def initialize(
    33 |              adapter:,
    34 |              ssl:,
    35 |              hostname:,
    36 |              port:,
    37 |              uds_path:,
    38 |              timeout_seconds:,
    39 |              deprecated_for_removal_transport_configuration_proc:
    40 |            )
    lib/datadog/core/configuration/agent_settings_resolver.rb:22: The corresponding argument adapter in the previous definition was not a keyword argument
    22 |        AgentSettings = \
    23 |          Struct.new(
    24 |            :adapter,
    25 |            :ssl,
    26 |            :hostname,
       |...
    48 |                deprecated_for_removal_transport_configuration_proc
    49 |              )
    50 |              freeze
    51 |            end
    52 |          end
Errors: 1
```

which in this case we don't want to do anything about.

That is because we're explicitly redefining the `Struct` with
keyword arguments because older Rubies don't have support for
keyword arguments in Structs.

Right now the only "fix" would be to not use a Struct (or not
use keyword arguments), which I don't think is worth it.