First step of making some sample values optional: Abstract values
A lot of the knowledge for the values was spread out everywhere --
the collectors had to care about a lot of libdatadog details that they
were getting from the `stack_recorder.h` header.

As part of the work to make some values optional, I've replaced
the previous `metric_values` array with a very concrete struct, and
moved the knowledge about how to turn it into the array inside the
`StackRecorder`.